requirements.txt
sqlalchemy
psycopg2-binary
requests
deep-translator
playwright
praw
jsonschema
tenacity
python-dotenv
fastapi
uvicorn
pyjwt
google-search-results
newsapi-python
unsplash-python